rename monster to enemy
